requirements.txt
absl-py==1.3.0
accelerate==0.16.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
apex==0.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blessed==1.19.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
chex==0.1.5
click==8.0.4
commonmark==0.9.1
contourpy==1.0.5
cryptography==37.0.1
cycler==0.11.0
datasets==2.10.1
debugpy==1.5.1
decorator==5.1.1
deepspeed==0.7.4
dgl==0.9.0
dill==0.3.5.1
dm-tree==0.1.7
docopt==0.6.2
einops==0.6.0
entrypoints==0.4
executing==0.8.3
expecttest==0.1.4
filelock==3.6.0
flatbuffers==22.11.23
flax==0.6.2
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.8.2
gast==0.4.0
gdown==4.6.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.0.0
grpcio==1.51.1
h5py==3.7.0
hjson==3.1.0
huggingface-hub==0.12.1
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.10.0
ipykernel==6.15.2
ipython==8.4.0
jax==0.3.25
jaxlib==0.3.25
jedi==0.18.1
joblib==1.1.0
jsonlines==3.1.0
jupyter_client==7.3.5
jupyter-core==4.10.0
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
nest-asyncio==1.5.5
networkx==2.8.4
ninja==1.11.1
nltk==3.8.1
numpy==1.21.5
nvidia-ml-py==11.495.46
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pipreqs==0.4.11
plotly==5.11.0
prompt-toolkit==3.0.20
protobuf==3.19.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.6.0
rouge==1.0.1
rsa==4.9
sacremoses==0.0.43
scikit-learn==1.1.2
scipy==1.7.3
sentencepiece==0.1.97
setuptools==63.4.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
style==1.1.0
SwissArmyTransformer==0.2.12
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
tensorstore==0.1.28
termcolor==2.1.1
threadpoolctl==3.1.0
tokenization==1.0.7
tokenizers==0.11.4
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1
torchmetrics==0.11.4
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
transformers==4.26.1
typing_extensions==4.3.0
undecorated==0.3.0
update==0.0.1
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
xxhash==3.0.0
yarg==0.1.9
yarl==1.8.1
zipp==3.8.0
